+++ package.json
@@ -82,6 +82,7 @@
     "babel-plugin-transform-runtime": "^6.7.5",
     "babel-polyfill": "^6.7.2",
     "babel-preset-react": "^6.5.0",
+    "classnames": "^2.2.5",
     "core-js": "^2.4.0",
     "css-loader": "^0.23.1",
     "cutestrap": "^1.3.1",
